requirements.txt
Django==4.1.3
python-dotenv
mysqlclient
django-filer
django-admin-sortable2
black
easy-thumbnails